Update insidedeclaration.bicep
I would assume more people are going to look for how to create blob containers over file shares with Bicep. I propose expanding on the sample to both illustrate how to use in-lined resources to define containers in addition to indicating that these in-lined resources can also be looped themselves.